+++ documentation/playground/src/organisms/FooterExamples.tsx
@@ -1,11 +1,12 @@
 import { Footer } from '@a-ui/react';
+import { Link } from 'packages/antwerp-ui/react-components/src/constants/application.types';
 
 export function FooterExamples() {
-  const items = [
-    { label: '© stad Antwerpen' },
-    { label: 'Privacy', href: '#' },
+  const items: Link[] = [
+    { label: '© stad Antwerpen', onClick: (event) => console.log(event) },
+    { label: 'Privacy' },
     { label: 'Toegankelijkheid', href: '#' },
-    { label: 'Cookie instellingen', href: '#' }
+    { label: 'Cookie instellingen', onClick: () => console.log('clicked') }
   ];
   return (
     <div className="u-margin">